Auto-load min sensor id when switching between different panes
